# DatingApp.API/openapi.yaml
openapi: 3.0.0
info:
  title: DatingApp API
  version: v1
  description: DatingApp.API
servers:
  - url: http://localhost:5000/api

components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned from POST /api/auth/login (without the "Bearer " prefix).

  schemas:
    Gender:
      type: string
      enum: [Man, Woman, Other]

    RegisterUserDto:
      type: object
      required: [firstName, lastName, email, password, confirmPassword, gender, interestedIn, dateOfBirth]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 64
        confirmPassword:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        interestedIn:
          $ref: '#/components/schemas/Gender'
        dateOfBirth:
          type: string
          format: date-time

    LoginUserDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    ListUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        gender:
          type: string
        age:
          type: integer
        knownAs:
          type: string
        created:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        city:
          type: string
        country:
          type: string
        photoUrl:
          type: string

    DetailedPhotoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        isMain:
          type: boolean

    DetailedUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        gender:
          type: string
        age:
          type: integer
        knownAs:
          type: string
        created:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        introduction:
          type: string
        lookingFor:
          type: string
        interests:
          type: string
        city:
          type: string
        country:
          type: string
        photoUrl:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/DetailedPhotoDto'

    ListLikeDto:
      type: object
      properties:
        likeId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/ListUserDto'

    ListMatchDto:
      type: object
      properties:
        matchId:
          type: string
          format: uuid
        matchedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/ListUserDto'

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string

    PagedListLikeDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ListLikeDto'
        nextCursor:
          type: string
          nullable: true
        pageSize:
          type: integer

    PagedListMatchDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ListMatchDto'
        nextCursor:
          type: string
          nullable: true
        pageSize:
          type: integer

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      description: >-
        gender/interestedIn must be one of: Man, Woman, Other.
        password is 8-64 chars. username is auto-generated server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
            example:
              firstName: Dana
              lastName: Lee
              email: dana@example.com
              password: password123
              confirmPassword: password123
              gender: Woman
              interestedIn: Man
              dateOfBirth: "1995-05-01T00:00:00Z"
      responses:
        '201':
          description: User registered successfully
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Login and receive a JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
            example:
              email: dana@example.com
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    get:
      summary: Get all users
      tags: [Users]
      security:
        - Bearer: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListUserDto'
        '401':
          description: Unauthorized

  /users/{id}:
    get:
      summary: Get a user by ID
      tags: [Users]
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detailed user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedUserDto'
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/ping:
    get:
      summary: Health-check stub for the Users controller
      tags: [Users]
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized

  /users/{id}/like/{recipientId}:
    post:
      summary: Like a user (creates a match if mutual)
      tags: [Likes]
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Calling user's ID (must match the authenticated user)
          schema:
            type: string
            format: uuid
        - name: recipientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Like created; isMatch indicates whether a match was formed
          content:
            application/json:
              schema:
                type: object
                properties:
                  isMatch:
                    type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (caller ID mismatch)
        '404':
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Already liked this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /likes:
    get:
      summary: Get users who liked the authenticated user (cursor-paginated)
      tags: [Likes]
      security:
        - Bearer: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor from the previous page's nextCursor
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated list of likes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedListLikeDto'
        '401':
          description: Unauthorized

  /matches:
    get:
      summary: Get matches for the authenticated user (cursor-paginated)
      tags: [Matches]
      security:
        - Bearer: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor from the previous page's nextCursor
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated list of matches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedListMatchDto'
        '401':
          description: Unauthorized

  /matches/{id}:
    delete:
      summary: Unmatch (delete a match)
      tags: [Matches]
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unmatched successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not a participant in this match)
        '404':
          description: Match not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'